Create new cli-group
