+++ BALSAMIC/commands/base.py
@@ -7,6 +7,7 @@ import coloredlogs
 
 # Subcommands
 from BALSAMIC.commands.run.base import run as run_command
+from BALSAMIC.commands.init.base import initialize as init_command
 from BALSAMIC.commands.report.base import report as report_command
 from BALSAMIC.commands.config.base import config as config_command
 from BALSAMIC.commands.plugins.base import plugins as plugins_command
@@ -67,3 +68,4 @@ cli.add_command(run_command)
 cli.add_command(report_command)
 cli.add_command(config_command)
 cli.add_command(plugins_command)
+cli.add_command(init_command)
